code refactored and ready is modified
